revised run_inspect()
Moved some global variables (FULLCONTENTS, TOPCONTENTS) not all of which
are used in run_inspect()) to parse_archive() and made them and
SOURCE_DIR derivable from one tar call. Moved the remaining default var
to the else clause. Shortened the perl 'grep' to an 'egrep'. Made the
'autotools' output consistent with the others. Used printf()'s field
width specifier in place of tabs and made the columns align on the
longest field rather than a large amount of whitespace. Changed the 'for
list' into a standard for..do..done. Made more ash-friendly.