unfunctioned get_commandline(); created validate_cfg()
This turned into a shotgun commit. It was meant to start revising the
variable assignments up top and 'main' down below.

Added foldmarker/dividers to the major sections and a modeline to go
with. Changed the {U,}TAG handling (which meant changing the conf file
and a line in common(), too) and moved that and the general conf file
sourcing to main, adding validate_cfg() to loosely help ensure a
coherent config. Broke get_commandline() out of its function and
inserted it near the top of main. Changed the 'exit's in
parse_archive(), version(), and check_version() to appropriate 'return's
and redirected check_version's error message to stderr. And exited if
parse_archive() returned non-zero. Converted main's 'if..then..fi'
one-liners into blocks.