renamed run_{build,inspect}() and the -r option
They are now build() and pick_template() and the -b|--build option. This
led to changing the RUN option to the _BUILD option (so as not to
conflict with the external SlackBuild 'BUILD' variable). Also made the
behavior and description of -i fully imply -b. And removed '-l' from the
README.