revised common() and the DOCS assignment
DOCS is now assigned to in a quoted alphabetical list with a simple
pre-assignment inclusion. The for loop in common() was made consistent
with the rest. All the '[[ -n' tests were changed to exit status tests.
The printf output was revised though it's functionally about the same. A
simple sed replaces the line number math for deleting unwanted parts.
Removed unnecessary variable braces. And the usual 'one sed' and 'seq
loop' and 'ash-friendly' and '80 column' changes.