underscore, quoting, tr, rm_man_info() issues
Fixing miscellaneous bugs and issues that more testing turned up.

After discussion with David, we decided to go with allowing (requiring)
underscores in TAG specifications to be consistent with SBo. Also thanks
to him for catching and fixing the 'tr' bug that took '-' in a package
name as a range and supplying the '[:graph:]' fix. Also fixed numerous
forgotten quotes in tests. Also used $(id -u) rather than $UID, which
isn't set in all shells. Finally, created rm_man_info() to make it
available to both autotools() and cmake().