fixed logging issue in build(); misc changes
Tar output from making a package from a SlackBuild was getting sent to
the error log rather than the display and now it goes to both (which may
not be ideal but should suffice).

Also addressed likely bugs in rm_special(), fix_build_errors(), and a
bashism in build(). Also removed the now-unused rm_man_info() and made
other minor tweaks.